Moved and updated tbb_lithe.h
Main purpose for the move was so that example application would have
access to this class when running.  Maybe it should be hidden from
external applications eventually, but for now it seems to work best
here.

Also, I updated the scheduler to allow heirarchical scheduling to work
properly.  Made me think a bit more about how lithe should actually be
designed.  may incorporate some of those ideas in soon.